Move the Pipeline and TaggerWrapper structs to the sticker crate
This data structures are too useful for downstream applications to
hide in the sticker-utils crate.

TaggerWrapper is also renamed to Tagger, since it is namespaced by the
wrapper module.
