+++ sticker-utils/src/lib.rs
@@ -1,25 +1,8 @@
-mod config;
-pub use crate::config::{
-    Config, Embedding, EmbeddingAlloc, Embeddings, EncoderType, Input, Labeler, LabelerType,
-};
-
-mod pipeline;
-pub use crate::pipeline::Pipeline;
-
 mod progress;
 pub use crate::progress::{ReadProgress, TaggerSpeed};
 
 mod sent_proc;
 pub use crate::sent_proc::SentProcessor;
-
-mod serialization;
-pub use crate::serialization::{CborRead, CborWrite, TomlRead};
-
-mod tagger_wrapper;
-pub use crate::tagger_wrapper::TaggerWrapper;
-
-#[cfg(test)]
-mod config_tests;
 
 mod save;
 pub use crate::save::{CompletedUnit, SaveSchedule, SaveScheduler};
