+++ sticker-utils/src/subcommands/tag.rs
@@ -3,7 +3,9 @@ use std::io::BufWriter;
 use clap::{App, Arg, ArgMatches};
 use conllx::io::{ReadSentence, Reader, WriteSentence, Writer};
 use stdinout::{Input, OrExit, Output};
-use sticker_utils::{Pipeline, SentProcessor, TaggerSpeed};
+use sticker::wrapper::Pipeline;
+
+use sticker_utils::{SentProcessor, TaggerSpeed};
 
 use crate::{StickerApp, StickerPipelineApp};
