+++ sticker/src/lib.rs
@@ -9,7 +9,11 @@ pub use crate::input::{Embeddings, InputVector, LayerEmbeddings, SentVectorizer}
 mod numberer;
 pub use crate::numberer::Numberer;
 
+pub mod serialization;
+
 mod tag;
 pub use crate::tag::{Layer, LayerValue, ModelPerformance, Tag, TopK, TopKLabels};
 
 pub mod tensorflow;
+
+pub mod wrapper;
