+++ sticker/testdata/sticker.conf
